Created Temp File
Sharing my attempts at getting the X's and O's to switch off, so we both have the code to try and get it to work. There are 2 different versions with a huge space between them. Things refused to work for me, and after a while I got totally lost, so I undid several steps to the last time anything changed, and that's what I have posted here.